sdk/python/requirements.txt
scrypt>=0.8.0
pycryptodome>=3.19.0
eth-abi>=4.2.0
web3>=6.11.0
